Add files via upload
V2: The objective of this code is to guess the correct predefined number, which is 12345. You can keep guessing until you get it right or choose to exit the game.

V1 groceries: The code is a straightforward, the user has a single opportunity to guess the correct number. It accurately takes input, validates it, and gives feedback. However, it could be enhanced by incorporating error handling and enabling multiple guesses."

IT5016: This code is a ticket management system for handling support tickets, including features for creating, resolving, and managing tickets.